+++ TS3AudioBot/TS3Query/Messages/ClientData.cs
@@ -3,19 +3,19 @@
 	// clientlist
 	public class ClientData : Response
 	{
-        [QuerySerialized("clid")]
-        public int Id;
+		[QuerySerialized("clid")]
+		public ushort Id;
 
-        [QuerySerialized("cid")]
-        public int ChannelId;
+		[QuerySerialized("cid")]
+		public int ChannelId;
 
-        [QuerySerialized("client_database_id")]
-        public int DatabaseId;
+		[QuerySerialized("client_database_id")]
+		public ulong DatabaseId;
 
-        [QuerySerialized("client_nickname")]
-        public string NickName;
+		[QuerySerialized("client_nickname")]
+		public string NickName;
 
-        [QuerySerialized("client_type")]
-        public ClientType Type;
-    }
+		[QuerySerialized("client_type")]
+		public ClientType Type;
+	}
 }
